removed draft toggle command in favor of enable/disable subcommand
